pyproject.toml
[build-system]
requires = [
    "setuptools>=60",
    "setuptools-scm>=8.0"]

[project]
name = "fm-factual"
dynamic = ["version"]
description = "Factuality for Foundation Models"
license = {text = "Apache-2.0"}
readme = "README.md"
requires-python = ">=3.9"
classifiers=[
    "License :: OSI Approved :: Apache Software License"
]
dependencies = [
    # TODO: loosen dependencies
    "accelerate>=0.22.0",
    "datasets>=2.4.0",
    "huggingface-hub",
    "numpy>=1.22.4",
    "pandas>=1.5.0",
    "scikit-learn>=1.1",
    "scipy>=1.8.1",
    "tokenizers>=0.13.3",
    "torch>=2.0.1",
    "tqdm>=4.65.0",
    "transformers>=4.32.0",
    "peft>=0.6.0",
    "evaluate>=0.4.1",
    "rouge>=1.0.1",
    "rouge_score>=0.1.2",
    "nltk>=3.6.6",
    "protobuf",
    "sentencepiece",
    "python-dotenv",
    "pgmpy",
    "matplotlib",
    "ipykernel",
    "ipywidgets",
    "pyomo",
    "rank_bm25",
    "spacy",
    "pysqlite3",
    "flask",
    "bert-score",
    "helper-utils",
    "langchain",
    "langchain-community",
    "wikipedia",
    "chromadb",
    "litellm",
    "beautifulsoup4",
    "lxml",
    "pypdf"
]

[tool.setuptools.packages.find]
exclude = ["tests", "tests.*"]
namespaces = false


[tool.setuptools_scm]
version_file = "fm_factual/version.py"

[project.urls]
Source = "https://github.ibm.com/IBM-Research-AI/fm-factual"